update: handle cache
